Update the expected parameter val to be uint64_t until it will be fixed on MC side to uint32_t

+++ src/lib/evmng.h
@@ -81,7 +81,7 @@ struct evmng_irq_params {
 	/** addr - address for the interrupt, provided by IRQ*/
 	uint64_t addr;
 	/** val - interrupt value provided via IRQ API*/
-	uint32_t val;
+	uint64_t val;
 };
 
 #endif /* __EVMNG_H */
